backend/requirements.txt
flask==3.0.0
requests==2.31.0
scikit-learn==1.3.0
nltk==3.8.1
sqlite3==2.6.0
google-generativeai==0.3.2